chore(core, structure): validation on references
