+++ packages/sanity/src/core/validation/index.ts
@@ -4,3 +4,7 @@ export {Rule} from './Rule'
 export type {ValidationContext} from './types'
 export {validateDocument, type ValidateDocumentOptions} from './validateDocument'
 export {validateDocumentObservable} from './validateDocument'
+export {
+  validateDocumentWithReferences,
+  type ValidationStatus,
+} from './validateDocumentWithReferences'
